feat: add cargo management
This update adds cargo management features by making cargo components listed on the cargo tab.  Additional fields for cargo are also added. The automatic calculation of weight and power is made optional (off by default).  The setting to activate auto calc is in the debug sub-window of the settings tab.  This default being off was chosen because there are some strong assumptions on how weight and power is calculated.  Please see wiki for details.