feat: add training notes field and progress indicator (#989)
This release adds a field ("Training Notes") to the skill sheet that helps keep track of training for a skill.  If a fraction (e.g. 1/5) is entered anywhere in the field, a progress indicator (pie chart) will show-up on the actor's skills list.  If a fraction can't be found, the chart defaults to 50%.