fix: add powered armor (#990)
This fix adds a checkbox to armor sheets to indicate the armor is powered and has no weight when worn.  Also, a bug is corrected for cases where multiple sets of armor is carried (quantity > 1) and the weight adjustment applied to all armor sets and not just the set worn.