chore: Handle Weblate merge conflict.
There was a merge problem of some sort on Weblate, to fix this I took the easy way out and got the git commits from weblate and applied them as a patch directly in twodsix.
After this is in I'll go to weblate and force it to update from the main git repo and ignore all that is not yet committed.
I handled the merge conflict by assuming the file with more text was right. :) Hopefully that won't mean too much is lost.
The relevant translators should probably check if it looks ok after this.

These are the commits from Weblate included in this one:

commit 207578e15b469f369fcfb8923ccd9f27ec6a6ff8 (HEAD -> master, weblate/master, origin/master, origin/HEAD)
Author: Alex <mithnar333@googlemail.com>
Date:   Mon Sep 4 07:03:38 2023 +0000

    Translated using Weblate (German)

    Currently translated at 75.8% (810 of 1068 strings)

    Translation: Twodsix - Cepheus & Traveller (Unofficial)/main
    Translate-URL: https://weblate.foundryvtt-hub.com/projects/twodsix/main/de/

commit 3684ab48a9e5cb695988c6b5c22de75a4a75cfb7
Author: Alex <mithnar333@googlemail.com>
Date:   Sun Sep 3 17:55:43 2023 +0000

    Translated using Weblate (German)

    Currently translated at 50.9% (544 of 1068 strings)

    Translation: Twodsix - Cepheus & Traveller (Unofficial)/main
    Translate-URL: https://weblate.foundryvtt-hub.com/projects/twodsix/main/de/

commit 8521703cd69dd3cf14506116b829e6aa59532fc4
Author: gitlocalize-app[bot] <55277160+gitlocalize-app[bot]@users.noreply.github.com>
Date:   Mon Sep 4 10:49:23 2023 -0500

    chore: Translate es.json via GitLocalize (#1459)

    Co-authored-by: ForjaSalvaje

# Conflicts:
#	static/lang/de.json