fix: Allow CUSTOM Active Effect calculations to use derived data (#1815)
This update reinstates the ability to use derived data with active effects using the CUSTOM calculation mode.  Please note that a side effect of this change is that using the CUSTOM override mode for an attribute will always take precedence over other calculation modes. Many thanks to Muji-k for the suggestions and help debugging.